backend/requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3